+++ src/005/problem005.rs
@@ -1,4 +1,4 @@
-fn greatest_common_divisor(a: i64, b: i64) -> i64 {
+fn greatest_common_divisor(a: i32, b: i32) -> i32 {
     let mut a = a;
     let mut b = b;
     while b != 0 {
@@ -9,11 +9,11 @@ fn greatest_common_divisor(a: i64, b: i64) -> i64 {
     a
 }
 
-fn least_common_multiple(a: i64, b: i64) -> i64 {
+fn least_common_multiple(a: i32, b: i32) -> i32 {
     a * b / greatest_common_divisor(a, b)
 }
 
-fn solve() -> i64 {
+fn solve() -> i32 {
     (1..20).fold(1, |acc, x| least_common_multiple(acc, x))
 }
